+++ src/main/java/com/wontlost/ckeditor/VaadinCKEditor.java
@@ -90,7 +90,7 @@ import static com.wontlost.ckeditor.JsonUtil.*;
  */
 @Tag("vaadin-ckeditor")
 @JsModule("./vaadin-ckeditor/vaadin-ckeditor.ts")
-@NpmPackage(value = "ckeditor5", version = "48.2.0")
+@NpmPackage(value = "ckeditor5", version = "48.4.0")
 @NpmPackage(value = "lit", version = "^3.3.3")
 public class VaadinCKEditor extends CustomField<String> implements HasAriaLabel {